Autentikasi dan Hak Akses

+++ Day1-API/routes/api.php
@@ -1,7 +1,10 @@
 <?php
 
 use Illuminate\Http\Request;
+// use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
+// use Spatie\Permission\Contracts\Role;
+// use Spatie\Permission\Models\Permission;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,20 +21,71 @@ use Illuminate\Support\Facades\Route;
 //     return $request->user();
 // });
 
-#Kategori
-Route::prefix('kategori')->group(function () {
-    Route::get('/', 'App\Http\Controllers\Api\KategoriController@index')->name('kategori.index');
-    Route::get('/{id}/show', 'App\Http\Controllers\Api\KategoriController@show')->name('kategori.show');
-    Route::post('/store', 'App\Http\Controllers\Api\KategoriController@store')->name('kategori.store');
-    Route::patch('/{id}/update', 'App\Http\Controllers\Api\KategoriController@update')->name('kategori.update');
-    Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\KategoriController@destroy')->name('kategori.destroy');
-});
+#login
+Route::post('/login', 'App\Http\Controllers\Api\LoginController@login')->name('login');
 
-#Produk
-Route::prefix('produk')->group(function () {
-    Route::get('/', 'App\Http\Controllers\Api\ProdukController@index')->name('produk.index');
-    Route::get('/{id}/show', 'App\Http\Controllers\Api\ProdukController@show')->name('produk.show');
-    Route::post('/store', 'App\Http\Controllers\Api\ProdukController@store')->name('produk.store');
-    Route::post('/{id}/update', 'App\Http\Controllers\Api\ProdukController@update')->name('produk.update');
-    Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\ProdukController@destroy')->name('produk.destroy');
+Route::middleware('auth:sanctum')->group(function () {
+    Route::prefix('kategori')->group(function () {
+        Route::get('/', 'App\Http\Controllers\Api\KategoriController@index')->name('kategori.index')->middleware('permission:lihat-kategori');
+        Route::get('/{id}/show', 'App\Http\Controllers\Api\KategoriController@show')->name('kategori.show')->middleware('permission:lihat-kategori');
+        Route::post('/store', 'App\Http\Controllers\Api\KategoriController@store')->name('kategori.store')->middleware('permission:tambah-kategori');
+        Route::post('/{id}/update', 'App\Http\Controllers\Api\KategoriController@update')->name('kategori.update')->middleware('permission:edit-kategori');
+        Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\KategoriController@destroy')->name('kategori.destroy')->middleware('permission:hapus-kategori');
+    });
+
+    Route::prefix('produk')->group(function () {
+        Route::get('/', 'App\Http\Controllers\Api\ProdukController@index')->name('produk.index')->middleware('permission:lihat-produk');
+        Route::get('/{id}/show', 'App\Http\Controllers\Api\ProdukController@show')->name('produk.show')->middleware('permission:lihat-produk');
+        Route::post('/store', 'App\Http\Controllers\Api\ProdukController@store')->name('produk.store')->middleware('permission:tambah-produk');
+        Route::post('/{id}/update', 'App\Http\Controllers\Api\ProdukController@update')->name('produk.update')->middleware('permission:edit-produk');
+        Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\ProdukController@destroy')->name('produk.destroy')->middleware('permission:hapus-produk');
+    });
 });
+// #Kategori
+// Route::prefix('kategori')->group(function () {
+//     Route::get('/', 'App\Http\Controllers\Api\KategoriController@index')->name('kategori.index');
+//     Route::get('/{id}/show', 'App\Http\Controllers\Api\KategoriController@show')->name('kategori.show');
+//     Route::post('/store', 'App\Http\Controllers\Api\KategoriController@store')->name('kategori.store');
+//     Route::patch('/{id}/update', 'App\Http\Controllers\Api\KategoriController@update')->name('kategori.update');
+//     Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\KategoriController@destroy')->name('kategori.destroy');
+// });
+
+// #Produk
+// Route::prefix('produk')->group(function () {
+//     Route::get('/', 'App\Http\Controllers\Api\ProdukController@index')->name('produk.index');
+//     Route::get('/{id}/show', 'App\Http\Controllers\Api\ProdukController@show')->name('produk.show');
+//     Route::post('/store', 'App\Http\Controllers\Api\ProdukController@store')->name('produk.store');
+//     Route::post('/{id}/update', 'App\Http\Controllers\Api\ProdukController@update')->name('produk.update');
+//     Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\ProdukController@destroy')->name('produk.destroy');
+// });
+
+
+
+// Route::group(['middleware' => ['auth:sanctum', 'role:admin']], function () {
+//     Route::prefix('kategori')->group(function () {
+//         Route::get('/', 'App\Http\Controllers\Api\KategoriController@index')->name('kategori.index')->middleware('permission:lihat-kategori');
+//         Route::get('/{id}/show', 'App\Http\Controllers\Api\KategoriController@show')->name('kategori.show')->middleware('permission:lihat-kategori');
+//         Route::post('/store', 'App\Http\Controllers\Api\KategoriController@store')->name('kategori.store')->middleware('permission:tambah-kategori');
+//         Route::patch('/{id}/update', 'App\Http\Controllers\Api\KategoriController@update')->name('kategori.update')->middleware('permission:edit-kategori');
+//         Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\KategoriController@destroy')->name('kategori.destroy')->middleware('permission:hapus-kategori');
+//     });
+//     Route::prefix('produk')->group(function () {
+//         Route::get('/', 'App\Http\Controllers\Api\ProdukController@index')->name('produk.index')->middleware('permission:lihat-produk');
+//     });
+// });
+
+
+
+//middleware untuk role penjual
+// Route::group(['middleware' => ['auth:sanctum', 'role:penjual']], function () {
+//     Route::prefix('kategori')->group(function () {
+//         Route::get('/', 'App\Http\Controllers\Api\KategoriController@index')->name('kategori.index');
+//     });
+//     Route::prefix('produk')->group(function () {
+//         Route::get('/', 'App\Http\Controllers\Api\ProdukController@index')->name('produk.index');
+//         Route::get('/{id}/show', 'App\Http\Controllers\Api\ProdukController@show')->name('produk.show');
+//         Route::post('/store', 'App\Http\Controllers\Api\ProdukController@store')->name('produk.store');
+//         Route::post('/{id}/update', 'App\Http\Controllers\Api\ProdukController@update')->name('produk.update');
+//         Route::delete('/{id}/destroy', 'App\Http\Controllers\Api\ProdukController@destroy')->name('produk.destroy');
+//     });
+// });
